Created seporate html and routes to create and edit courses
separated functions in portal.py to make the code cleaner and easier to read

changed css to make the p[ages slightly easier to read

debugged previous merge problems from having to many forms on a single page
