+++ portal/templates/layouts/courses/edit.html
@@ -4,14 +4,12 @@
 <section class="edit">
 <h1>Edit course</h1>
 <form class="teacher_home_form" method="post">
-  <label for="major">What Major does the new course belong to?
-  <input type="major" name="major" value="{{ course['major']}}" required/>
-  </label>
-  <label for="new_course">What will be the new course's name?</label>
+
+  <label for="new_course">What will be the course's new name?</label>
     <input type="new_course" name="new_course" value="{{ course['name']}}" required />
     <span>Course Description:</span>
     <textarea class="course-description" name="course_description">{{ course['description']}}</textarea>
-    <input type="submit" value="Create New Course"/>
+    <input type="submit" value="Submit Changes"/>
   </form>
 </section>
 {% endblock %}
